1.更新优化

+++ app/src/main/java/com/huacheng/huiservers/utils/ucrop/ImgCropUtil.java
@@ -231,11 +231,13 @@ public class ImgCropUtil {
             public void run() {
                 if (file.isDirectory()) {
                     File[] files = file.listFiles();
-                    for (int i = 0; i < files.length; i++) {
-                        File f = files[i];
-                        deleteFile(f);
+                    if (files!=null&&files.length>0){
+                        for (int i = 0; i < files.length; i++) {
+                            File f = files[i];
+                            deleteFile(f);
+                        }
+                        //    file.delete();//如要保留文件夹，只删除文件，请注释这行
                     }
-                    //    file.delete();//如要保留文件夹，只删除文件，请注释这行
                 }else if (file.exists()) {
                     file.delete();
                 }
